+++ src/history/history_utils.c
@@ -6,16 +6,16 @@
 /*   By: jumanner <jumanner@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 11:39:01 by jumanner          #+#    #+#             */
-/*   Updated: 2023/02/02 13:44:56 by jumanner         ###   ########.fr       */
+/*   Updated: 2023/02/02 14:39:03 by jumanner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "history.h"
 
-size_t	history_seek(t_state *state, size_t start, char *look_for,
+int	history_seek(t_state *state, int start, char *look_for,
 					bool anywhere)
 {
-	size_t	i;
+	int	i;
 
 	i = start;
 	while (i < HISTORY_SIZE)
